+++ R/stop_words.R
@@ -17,18 +17,19 @@
 #' }
 "stop_words"
 
-#' Get a tidy data frame of a single quanteda stop word lexicon
+#' Get a tidy data frame of a single stopword lexicon
 #'
-#' Get a specific stop word lexicon from quanteda's \link[quanteda]{stopwords}
-#' in a tidy format, with one row per word, in a form that can be joined or
-#' anti-joined with a one-word-per-row dataset.
+#' Get a specific stop word lexicon via the \pkg{stopwords} packages's
+#' \link[stopwords]{stopwords} function, in a tidy format with one word per row.
 #'
-#' @param language The language of the stop word lexicon specified, one of
-#' \code{"english"}, \code{"SMART"}, \code{"danish"}, \code{"french"},
-#' \code{"greek"}, \code{"hungarian"}, \code{"norwegian"}, \code{"russian"},
-#' \code{"swedish"}, \code{"catalan"}, \code{"dutch"}, \code{"finnish"},
-#' \code{"german"}, \code{"italian"}, \code{"portuguese"}, \code{"spanish"},
-#' or \code{"arabic"}
+#' @param language The language of the stopword lexicon specified as a
+#' two-letter ISO code, such as \code{"es"}, \code{"de"}, or \code{"fr"}.
+#' Default is \code{"en"} for English. Use
+#' \link[stopwords]{stopwords_getlanguages} from \pkg{stopwords} to see available
+#' languages.
+#' @param source The source of the stopword lexicon specified. Default is
+#' \code{"snowball"}. Use \link[stopwords]{stopwords_getsources} from
+#' \pkg{stopwords} to see available sources.
 #'
 #' @return A tibble with two columns, \code{word} and \code{lexicon}. The
 #' parameter \code{lexicon} is "quanteda" in this case.
@@ -36,12 +37,15 @@
 #' @examples
 #'
 #' library(dplyr)
-#' get_quanteda_stopwords("spanish")
-#' get_quanteda_stopwords("arabic")
+#' get_stopwords()
+#' get_stopwords(source = "smart")
+#' get_stopwords("es", "snowball")
+#' get_stopwords("ru", "snowball")
 #'
 #' @export
 #'
-get_quanteda_stopwords <- function(language = "english") {
-  data_frame(word = quanteda::stopwords(kind = language), lexicon = "quanteda")
+get_stopwords <- function(language = "en", source = "snowball") {
+  data_frame(word = stopwords::stopwords(language = language, source = source),
+             lexicon = source)
 }