+++ tests/testthat/test-stop-words.R
@@ -2,15 +2,15 @@ context("Stop words")
 
 suppressPackageStartupMessages(library(dplyr))
 
-test_that("get_quanteda_stopwords works for multiple languages", {
+test_that("get_stopwords works for multiple languages", {
 
-  skip_if_not_installed("quanteda")
-  de <- get_quanteda_stopwords("german")
-  ru <- get_quanteda_stopwords("russian")
+  skip_if_not_installed("stopwords")
+  de <- get_stopwords("de")
+  ru <- get_stopwords("ru")
 
   expect_is(de, "tbl_df")
   expect_is(ru, "tbl_df")
   expect_gt(nrow(de), nrow(ru))
-  expect_equal(unique(de$lexicon), "quanteda")
+  expect_equal(unique(de$lexicon), "snowball")
 
 })